Right now it can fire bullets. and moved the camera from the scene to playerInfol.

+++ NYP_Framework/Base/Source/WeaponInfo/Weapon.h
@@ -17,7 +17,7 @@ public:
 
 protected:
     unsigned char maxBullets, currBullets, maxClips, currClips;
-    double timeCounter, fireRate;
+    float timeCounter, fireRate;
 
 private:
     static size_t zeID;
